refactor(connect): split gateway message handlers
