+++ pkg/connect/gateway.go
@@ -29,7 +29,6 @@ import (
 	"github.com/oklog/ulid/v2"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 const (
@@ -785,418 +784,6 @@ func (c *connectGatewaySvc) Handler() http.Handler {
 			return
 		}
 	})
-}
-
-func (c *connectionHandler) handleIncomingWebSocketMessage(msg *connectpb.ConnectMessage) *connecterrors.SocketError {
-	c.log.Trace("received WebSocket message", "kind", msg.Kind.String())
-
-	switch msg.Kind {
-	case connectpb.GatewayMessageType_WORKER_READY:
-		// Do not allow marking worker as ready when gateway is draining
-		if c.svc.isDraining.Load() {
-			c.log.Warn("ignoring worker ready as svc is draining",
-				"instance_id", c.conn.Data.InstanceId,
-				"env_id", c.conn.EnvID.String(),
-				"account_id", c.conn.AccountID.String(),
-				"gateway_id", c.conn.GatewayId.String(),
-				"connection_id", c.conn.ConnectionId.String(),
-				"conn_draining", c.draining.Load(),
-			)
-
-			return &ErrDraining
-		}
-
-		// Check if connection itself is already marked as draining
-		if c.draining.Load() {
-			c.log.Warn("ignoring worker ready as connection is marked as draining")
-
-			// Connection is draining, so don't reset status to READY.
-			return nil
-		}
-
-		err := c.updateConnStatus(connectpb.ConnectionStatus_READY, "worker ready message")
-		if err != nil {
-			// TODO Should we actually close the connection here?
-			return &connecterrors.SocketError{
-				SysCode:    syscode.CodeConnectInternal,
-				StatusCode: websocket.StatusInternalError,
-				Msg:        "could not update connection status",
-			}
-		}
-
-		// Refresh worker capacity TTL if set
-		if c.conn.Data.InstanceId == "" {
-			// No instance ID, no capacity TTL to refresh
-			return &connecterrors.SocketError{
-				SysCode:    syscode.CodeConnectInternal,
-				StatusCode: websocket.StatusInternalError,
-				Msg:        "missing instanceId for connect message",
-			}
-		}
-
-		// Refresh worker capacity TTL if capacity is set
-		if err := c.svc.stateManager.WorkerCapacityOnHeartbeat(context.Background(), c.conn.EnvID, c.conn.Data.InstanceId); err != nil {
-			// Log but don't fail the heartbeat if TTL refresh fails
-			c.log.ReportError(err, "failed to refresh worker capacity TTL on heartbeat",
-				logger.WithErrorReportTags(map[string]string{
-					"instance_id":   util.SanitizeLogField(c.conn.Data.InstanceId),
-					"env_id":        c.conn.EnvID.String(),
-					"account_id":    c.conn.AccountID.String(),
-					"gateway_id":    c.conn.GatewayId.String(),
-					"connection_id": c.conn.ConnectionId.String(),
-				}))
-		}
-
-		for _, l := range c.svc.lifecycles {
-			go l.OnReady(context.Background(), c.conn)
-		}
-
-		c.log.Debug("marked worker as ready")
-
-		return nil
-	case connectpb.GatewayMessageType_WORKER_HEARTBEAT:
-		// Always allow heartbeats
-
-		// Don't reset status to READY if the connection or service is draining.
-		status := connectpb.ConnectionStatus_READY
-		if c.svc.isDraining.Load() || c.draining.Load() {
-			status = connectpb.ConnectionStatus_DRAINING
-
-			c.log.Warn("worker heartbeat received during draining sequence",
-				"conn_draining", c.draining.Load(),
-				"svc_draining", c.svc.isDraining.Load(),
-			)
-		}
-
-		err := c.updateConnStatus(status, "worker heartbeat",
-			"conn_draining", c.draining.Load(),
-			"svc_draining", c.svc.isDraining.Load(),
-		)
-		if err != nil {
-
-			c.log.ReportError(err, "failed to update connection status after heartbeat")
-
-			// TODO Should we actually close the connection here?
-			return &connecterrors.SocketError{
-				SysCode:    syscode.CodeConnectInternal,
-				StatusCode: websocket.StatusInternalError,
-				Msg:        "could not update connection status",
-			}
-		}
-
-		// Refresh worker capacity TTL if set
-		if c.conn.Data.InstanceId == "" {
-			// No instance ID, no capacity TTL to refresh
-			return &connecterrors.SocketError{
-				SysCode:    syscode.CodeConnectInternal,
-				StatusCode: websocket.StatusInternalError,
-				Msg:        "missing instanceId for connect message",
-			}
-		}
-
-		// Refresh worker capacity TTL if capacity is set
-		if err := c.svc.stateManager.WorkerCapacityOnHeartbeat(context.Background(), c.conn.EnvID, c.conn.Data.InstanceId); err != nil {
-			// Log but don't fail the heartbeat if TTL refresh fails
-			c.log.ReportError(err, "failed to refresh worker capacity TTL on heartbeat",
-				logger.WithErrorReportTags(map[string]string{
-					"instance_id":   util.SanitizeLogField(c.conn.Data.InstanceId),
-					"env_id":        c.conn.EnvID.String(),
-					"account_id":    c.conn.AccountID.String(),
-					"gateway_id":    c.conn.GatewayId.String(),
-					"connection_id": c.conn.ConnectionId.String(),
-				}))
-		}
-
-		for _, l := range c.svc.lifecycles {
-			go l.OnHeartbeat(context.Background(), c.conn)
-		}
-
-		// Respond with gateway heartbeat
-		writeCtx, writeCancel := context.WithTimeout(context.Background(), wsWriteTimeout)
-		defer writeCancel()
-		if err := wsproto.Write(writeCtx, c.ws, &connectpb.ConnectMessage{
-			Kind: connectpb.GatewayMessageType_GATEWAY_HEARTBEAT,
-		}); err != nil {
-			// The connection will fail to read and be closed in the read loop
-			return nil
-		}
-
-		c.setLastHeartbeat(time.Now())
-		c.log.Trace("worker heartbeat processed", "status", status.String())
-
-		return nil
-	case connectpb.GatewayMessageType_WORKER_STATUS:
-		if time.Since(c.getLastStatus()) < 2*time.Second {
-			c.log.Trace("ignoring WORKER_STATUS, rate limited")
-			return nil
-		}
-		c.setLastStatus(time.Now())
-
-		var data connectpb.WorkerStatusData
-		if err := proto.Unmarshal(msg.Payload, &data); err != nil {
-			c.log.Warn("invalid WORKER_STATUS payload", "err", err)
-			return nil
-		}
-
-		c.log.Debug("worker status received",
-			"in_flight_request_count", len(data.InFlightRequestIds),
-			"in_flight_request_ids", data.InFlightRequestIds,
-			"shutdown_requested", data.ShutdownRequested,
-		)
-		return nil
-	case connectpb.GatewayMessageType_WORKER_PAUSE:
-		// NOTE: Unlike WORKER_READY, we intentionally do
-		// NOT reject WORKER_PAUSE when the gateway is draining. The worker is
-		// signaling that it wants to stop receiving new requests (e.g. graceful
-		// shutdown via SIGTERM). We must honour this by updating the connection
-		// status in Redis to DRAINING so the router stops selecting it.
-		if c.svc.isDraining.Load() {
-			c.log.Warn("worker pause signal received during draining sequence")
-		}
-
-		// Mark as draining before updating Redis so that concurrent heartbeat
-		// messages do not reset the status back to READY.
-		c.draining.Store(true)
-
-		// Update the Redis status to DRAINING so the router stops selecting
-		// this connection for new requests.
-		err := c.updateConnStatus(connectpb.ConnectionStatus_DRAINING, "worker pause message", "svc_draining", c.svc.isDraining.Load())
-		if err != nil {
-			c.log.Error("could not update connection status to DRAINING on WORKER_PAUSE",
-				"err", err,
-			)
-		}
-
-		// Remove from in-memory map to prevent new requests from being forwarded.
-		c.svc.wsConnections.Delete(c.conn.ConnectionId.String())
-		c.stopForwardingOnce.Do(func() { close(c.stopForwarding) })
-
-		for _, l := range c.svc.lifecycles {
-			go l.OnStartDraining(context.Background(), c.conn)
-		}
-
-		return nil
-	case connectpb.GatewayMessageType_WORKER_REQUEST_ACK:
-		{
-			var data connectpb.WorkerRequestAckData
-			if err := proto.Unmarshal(msg.Payload, &data); err != nil {
-				// This should never happen: Failing the ack means we will redeliver the same request even though
-				// the worker already started processing it.
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectWorkerRequestAckInvalidPayload,
-					StatusCode: websocket.StatusPolicyViolation,
-					Msg:        "invalid payload in worker request ack",
-				}
-			}
-
-			// Unblock the Forward() call that's waiting for this ACK.
-			if ackCh, ok := c.pendingAcks.LoadAndDelete(data.RequestId); ok {
-				close(ackCh.(chan struct{}))
-			} else {
-				c.log.Warn(
-					"worker request ack received for unknown request ID",
-					"conn_id", c.conn.ConnectionId.String(),
-					"req_id", data.RequestId,
-					"run_id", data.RunId,
-				)
-			}
-
-			// This will be sent exactly once, as the router selected this gateway to handle the request
-			// Even if the gateway is draining, we should ack the message, the SDK will buffer messages and use a new connection to report results
-			grpcClient, err := c.svc.getOrCreateGRPCClient(context.Background(), c.conn.EnvID, data.RequestId)
-			if err != nil {
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectInternal,
-					StatusCode: websocket.StatusInternalError,
-					Msg:        "could not create grpc client to ack",
-				}
-			}
-
-			reply, err := grpcClient.Ack(context.Background(), &connectpb.AckMessage{
-				RequestId: data.RequestId,
-				Ts:        timestamppb.Now(),
-			})
-			if err != nil {
-				// This should never happen: Failing the ack means we will redeliver the same request even though
-				// the worker already started processing it.
-				c.log.ReportError(err, "failed to ack message through gRPC")
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectInternal,
-					StatusCode: websocket.StatusInternalError,
-					Msg:        "could not ack message through gRPC",
-				}
-			}
-			if !reply.Success {
-				c.log.Warn("failed to ack, executor was likely done with the request")
-			}
-			c.log.Trace("worker acked message",
-				"req_id", data.RequestId,
-				"run_id", data.RunId,
-				"transport", "grpc",
-			)
-
-			// TODO Should we send a reverse ack to the worker to start processing the request?
-
-			return nil
-		}
-	case connectpb.GatewayMessageType_WORKER_REPLY:
-		// Always handle SDK reply, even if gateway is draining
-		err := c.handleSdkReply(context.Background(), msg)
-		if err != nil {
-			c.log.ReportError(err, "could not handle sdk reply")
-			// TODO Should we actually close the connection here?
-			return &connecterrors.SocketError{
-				SysCode:    syscode.CodeConnectInternal,
-				StatusCode: websocket.StatusInternalError,
-				Msg:        "could not handle SDK reply",
-			}
-		}
-	case connectpb.GatewayMessageType_WORKER_REQUEST_EXTEND_LEASE:
-		{
-			var data connectpb.WorkerRequestExtendLeaseData
-			if err := proto.Unmarshal(msg.Payload, &data); err != nil {
-				// This should never happen: Failing the ack means we will redeliver the same request even though
-				// the worker already started processing it.
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectWorkerRequestExtendLeaseInvalidPayload,
-					StatusCode: websocket.StatusPolicyViolation,
-					Msg:        "invalid payload in worker request extend lease",
-				}
-			}
-
-			leaseID, err := ulid.Parse(data.LeaseId)
-			if err != nil {
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectWorkerRequestExtendLeaseInvalidPayload,
-					StatusCode: websocket.StatusPolicyViolation,
-					Msg:        "invalid lease ID in worker request extend lease payload",
-				}
-			}
-
-			// get worker capacity to check if we have worker limits to enforce
-			workerCap, err := c.svc.stateManager.GetWorkerCapacities(context.Background(), c.conn.EnvID, c.conn.Data.InstanceId)
-			if err != nil {
-				c.log.ReportError(err, "failed to get worker available capacity",
-					logger.WithErrorReportTags(map[string]string{
-						"instance_id": c.conn.Data.InstanceId,
-						"env_id":      c.conn.EnvID.String(),
-					}))
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectInternal,
-					StatusCode: websocket.StatusInternalError,
-					Msg:        "failed to get total worker capacity",
-				}
-			}
-			c.log.Trace("worker capacity info before extending lease", "account_id", c.conn.AccountID, "env_id", c.conn.EnvID, "instance_id", c.conn.Data.InstanceId, "worker_total_capacity", workerCap.Total, "worker_available_capacity", workerCap.Available)
-
-			// extend lease with worker capacity limit if set
-			newLeaseID, err := c.svc.stateManager.ExtendRequestLease(context.Background(), c.conn.EnvID, c.conn.Data.InstanceId,
-				data.RequestId, leaseID, consts.ConnectWorkerRequestLeaseDuration, workerCap.IsUnlimited())
-			if err != nil {
-				switch {
-				case errors.Is(err, state.ErrRequestLeaseExpired),
-					errors.Is(err, state.ErrRequestLeased),
-					errors.Is(err, state.ErrRequestLeaseNotFound),
-					errors.Is(err, state.ErrRequestWorkerDoesNotExist):
-
-					c.log.ReportError(err, "lease was claimed by other worker, expired, or worker does not exist",
-						logger.WithErrorReportTags(map[string]string{
-							"req_id":   data.RequestId,
-							"lease_id": leaseID.String(),
-						}))
-
-					// Respond with nack
-					nackPayload, marshalErr := proto.Marshal(&connectpb.WorkerRequestExtendLeaseAckData{
-						RequestId:    data.RequestId,
-						AccountId:    data.AccountId,
-						EnvId:        data.EnvId,
-						AppId:        data.AppId,
-						FunctionSlug: data.FunctionSlug,
-
-						// No new lease ID
-						NewLeaseId: nil,
-					})
-					if marshalErr != nil {
-						// This should never happen
-						return &connecterrors.SocketError{
-							SysCode:    syscode.CodeConnectInternal,
-							StatusCode: websocket.StatusInternalError,
-							Msg:        "failed to marshal nack payload",
-						}
-					}
-
-					nackWriteCtx, nackWriteCancel := context.WithTimeout(context.Background(), wsWriteTimeout)
-					defer nackWriteCancel()
-					if err := wsproto.Write(nackWriteCtx, c.ws, &connectpb.ConnectMessage{
-						Kind:    connectpb.GatewayMessageType_WORKER_REQUEST_EXTEND_LEASE_ACK,
-						Payload: nackPayload,
-					}); err != nil {
-						// The connection will fail to read and be closed in the read loop
-						return nil
-					}
-
-					return nil
-
-				default:
-					c.log.ReportError(err, "unexpected error extending lease",
-						logger.WithErrorReportTags(map[string]string{
-							"req_id":   data.RequestId,
-							"lease_id": leaseID.String(),
-						}))
-
-					// This should never happen
-					return &connecterrors.SocketError{
-						SysCode:    syscode.CodeConnectInternal,
-						StatusCode: websocket.StatusInternalError,
-						Msg:        "unexpected error extending lease",
-					}
-				}
-			}
-
-			var newLeaseIDStr *string
-			if newLeaseID != nil {
-				newLeaseIDStr = proto.String(newLeaseID.String())
-			}
-
-			// Respond with ack
-			ackPayload, marshalErr := proto.Marshal(&connectpb.WorkerRequestExtendLeaseAckData{
-				RequestId:    data.RequestId,
-				AccountId:    data.AccountId,
-				EnvId:        data.EnvId,
-				AppId:        data.AppId,
-				FunctionSlug: data.FunctionSlug,
-				NewLeaseId:   newLeaseIDStr,
-			})
-			if marshalErr != nil {
-				// This should never happen
-				return &connecterrors.SocketError{
-					SysCode:    syscode.CodeConnectInternal,
-					StatusCode: websocket.StatusInternalError,
-					Msg:        "failed to marshal nack payload",
-				}
-			}
-
-			ackWriteCtx, ackWriteCancel := context.WithTimeout(context.Background(), wsWriteTimeout)
-			defer ackWriteCancel()
-			if err := wsproto.Write(ackWriteCtx, c.ws, &connectpb.ConnectMessage{
-				Kind:    connectpb.GatewayMessageType_WORKER_REQUEST_EXTEND_LEASE_ACK,
-				Payload: ackPayload,
-			}); err != nil {
-				// The connection will fail to read and be closed in the read loop
-				return nil
-			}
-
-			c.log.Debug("extended lease for long-running request", "req_id", data.RequestId)
-
-			// Extended lease, all good
-			return nil
-		}
-	default:
-		c.log.Warn("unexpected message kind received", "kind", msg.Kind.String())
-		return nil
-	}
-
-	return nil
 }
 
 func (c *connectionHandler) receiveRouterMessagesFromGRPC(ctx context.Context, onSubscribed chan struct{}) {
